Fix issue that prevented fields from being unset.
Since it only did the update when value is not None, it was impossible to
override a value already in the api_data dict with an explicit None.

Now the Nones will overwrite what was in api_data and we'll clean up any
explicit Nones that are left in the dict after the fact, preserving the
previous behavior that we never explicitly send a JSON null, instead
preferring to just omit null properties.